requirements.txt
selene==2.0.0rc9
pytest==8.4.1
allure-pytest
selenium==4.12.0